refactro: use one tile component for assets
