+++ packages/desktop/views/dashboard/send-flow/views/components/index.js
@@ -1,5 +1,4 @@
 export { default as DateTimePicker } from './DateTimePicker.svelte'
 export { default as DateTimePickerButton } from './DateTimePickerButton.svelte'
 export { default as DateTimePickerMenu } from './DateTimePickerMenu.svelte'
-export { default as TokenAmountTile } from './TokenAmountTile.svelte'
 export { default as TransactionDetails } from './TransactionDetails.svelte'
