+++ packages/shared/components/atoms/tiles/ClickableTile.svelte
@@ -3,13 +3,13 @@
 
     export let disabled = false
     export let classes = ''
-    export let onClick: () => void
+    export let onClick: (() => void) | undefined = undefined
 </script>
 
 <Tile
     classes="
-        hover:bg-gray-100 dark:hover:bg-gray-1000
-        {disabled ? 'opacity-50 pointer-events-none' : 'cursor-pointer'}
+        {onClick ? 'hover:bg-gray-100 dark:hover:bg-gray-1000 cursor-pointer' : 'pointer-events-none'}
+        {disabled ? 'opacity-50' : ''}
         {classes}
     "
     on:click={onClick}
