+++ packages/shared/components/atoms/tiles/index.js
@@ -2,8 +2,8 @@ export * from './tileContents'
 export * from './tileFooters'
 
 export { default as ActivityTile } from './ActivityTile.svelte'
-export { default as AssetTile } from './AssetTile.svelte'
 export { default as TokenAmountTile } from './TokenAmountTile.svelte'
+export { default as TokenAvailableBalanceTile } from './TokenAvailableBalanceTile.svelte'
 export { default as ClickableTile } from './ClickableTile.svelte'
 export { default as ShimmerClaimingAccountTile } from './ShimmerClaimingAccountTile.svelte'
 export { default as Tile } from './Tile.svelte'
